new task 7-1 17.12

+++ lab-7/task 7-1.sql
@@ -3,10 +3,10 @@ USE cd;
 DELIMITER $$
 DROP FUNCTION IF EXISTS CalculateRentalCost;
 CREATE FUNCTION CalculateRentalCost(memid INT, facid INT, slots INT)
-RETURNS INT
+RETURNS DECIMAL(10, 2)
 DETERMINISTIC
 BEGIN
-	DECLARE cost INT;
+	DECLARE cost DECIMAL(10, 2);
 	SET cost = (SELECT IF(memid = 0, guestcost * slots, membercost * slots)
 	FROM facilities f WHERE facid = f.facid);
 	RETURN cost;
